start of create idea form
