--- src/components/index.ts
@@ -1,18 +1,0 @@
-export * from './common/button';
-export * from './common/tabs';
-export * from './common/link';
-export * from './common/icon';
-export * from './common/tag';
-export * from './common/modal';
-export * from './common/input';
-export * from './common/textarea';
-export * from './common/section-header';
-
-export * from './feature-specific/idea-card';
-export * from './feature-specific/solution-card';
-export * from './feature-specific/overall-card';
-
-export * from './page-specific/home/overall-section';
-
-export * from './demo-wrapper';
-export * from './example-with-tabs';
